refactor(eval): extract scoreTrigger/scoreOutput into pure score-lib (behaviour-preserving, golden-tested)
