+++ tests/fixtures/eval/output.golden.md
@@ -1,0 +1,16 @@
+# Output eval scorecard
+
+Skills evaluated: 2. Output checks: 4.
+
+**Overall: 75% of checks passed** (3/4). Skills passing every check: 1/2.
+
+| Skill | checks passed | artifact chars |
+|---|---|---|
+| premortem | 100% (2/2) | 1234 |
+| swot | 50% (1/2) | 999 |
+
+## Failed checks (1)
+
+**swot** (1/2)
+- FAIL: "d2" - missing caveat
+
